refactor image match
