+++ datgen/input_prepro/object.py
@@ -12,7 +12,7 @@ class DGObject():
         self.obj_name = vals['obj'].lower()
         self.vis_attr = [a.lower() for a in vals['vis_attr']]
         self.loc = [l.lower() for l in vals['loc']]
-        self.n_imgs = vals['n_images']
+        self.n_images = vals['n_images']
         self.size_min = vals['size_min']
         
         # Generate object-attribute combination
@@ -109,17 +109,3 @@ def _clean_captions(captions):
         caption = caption.capitalize() + '.'
         clean_captions.append(caption)
     return clean_captions
-
-
-# def create_property_list(inputs):
-#     for obj, vals in inputs.items():
-#         inputs[obj]['obj'] = vals['obj'].lower()
-#         inputs[obj]['vis_attr'] = [a.lower() for a in vals['vis_attr']]
-#         inputs[obj]['loc'] = [l.lower() for l in vals['loc']]
-#         if vals['vis_attr'] != ['']:
-#             inputs[obj]['obj_attr'] = [
-#                 f'{a} {inputs[obj]["obj"]}' for a in inputs[obj]['vis_attr']
-#             ]
-#         else:
-#             inputs[obj]['obj_attr'] = ['']
-#     return inputs
